refactor(controls): delete the dead view/controls render path
With every Settings control rendering through the widget framework,
the view/controls ratatui renderers had no callers left. Delete the
render modules (toggle, number_input, dropdown, text_input, dual_list,
text_list, map_input, keybinding_list, button) and their render-only
tests + test_frame helpers; slim the module re-exports to the model
types (`*State`, events, and the `DualListLayout` the settings layout
still names). `format_key_combo` — a domain formatter the widget
mapping uses, not a renderer — moves into keybinding_list/mod.rs.

There is now exactly one renderer per control: the plugin widget
framework. The `*State` structs remain as the settings model until
input routes through the widget runtime.

Verified: lib 2944/2944 (render-only tests removed), settings e2e
filter 141/141 (4 ignored).
